+++ battlecode22-scaffold/src/OFinalSprintTwoBot/BFSDroid.java
@@ -1,4 +1,4 @@
-package AFinalSprintTwoBot;
+package OFinalSprintTwoBot;
 
 import battlecode.common.Direction;
 import battlecode.common.MapLocation;