+++ battlecode22-scaffold/src/OFinalSprintTwoBot/CombatUtil.java
@@ -1,4 +1,4 @@
-package AFinalSprintTwoBot;
+package OFinalSprintTwoBot;
 
 import battlecode.common.*;